display user profile - display form to edit user details

+++ src/main/java/fr/poei/fines_saveurs_fo/service/CustomerService.java
@@ -24,7 +24,7 @@ public class CustomerService {
         return customerRepository.findAll();
     }
 
-    public Optional<Customer> fetchById(int id) {
+    public Optional<Customer> fetchById(long id) {
         return customerRepository.findById(id);
     }
